fix: Rebase conflicts and cleanup
