+++ src/oogeso/dto/serialisation.py
@@ -1,12 +1,16 @@
 import json
 import logging
-from dataclasses import asdict, is_dataclass
 from typing import Any, Dict, List, Optional, Union
 
 import pandas as pd
+from pydantic import BaseModel
 
 from oogeso import dto
-from oogeso.utils.util import get_class_from_dto
+from oogeso.dto.mapper import (
+    get_carrier_data_class_from_str,
+    get_device_data_class_from_str,
+    get_edge_data_class_from_str,
+)
 
 logger = logging.getLogger(__name__)
 
@@ -15,15 +19,16 @@ class DataclassJSONEncoder(json.JSONEncoder):
     """Serialize (save to file)"""
 
     def default(self, obj: Any):
-        if is_dataclass(obj=obj):
-            dct = asdict(obj=obj)
-            return dct
+        if isinstance(obj, BaseModel):
+            return obj.dict()
         if isinstance(obj, pd.Series):
             dct = obj.reset_index().to_json(orient="records")
             return dct
         if isinstance(obj, pd.DataFrame):
             dct = obj.reset_index().to_json(orient="records")
             return dct
+        if isinstance(obj, BaseModel):
+            return obj.dict()
         return super().default(obj)
 
 
@@ -36,8 +41,7 @@ class DataclassJSONDecoder(json.JSONDecoder):
     @staticmethod
     def _new_carrier(dct: Dict[str, str]):
         model = dct["id"]
-        carrier_class_str = f"Carrier{model.capitalize()}Data"
-        carrier_class = get_class_from_dto(class_str=carrier_class_str)
+        carrier_class = get_carrier_data_class_from_str(model_name=model)
         return carrier_class(**dct)
 
     @staticmethod
@@ -47,20 +51,19 @@ class DataclassJSONDecoder(json.JSONDecoder):
     @staticmethod
     def _new_edge(dct: Dict[str, str]) -> dto.EdgeData:
         carrier = dct.pop("carrier")  # gets and deletes model from dictionary
-        edge_class_str = f"Edge{carrier.capitalize()}Data"
-        edge_class = get_class_from_dto(class_str=edge_class_str)
+        edge_class = get_edge_data_class_from_str(carrier_name=carrier)
         return edge_class(**dct)
 
     @staticmethod
     def _new_device(dct: Dict[str, Union[str, object]]) -> dto.DeviceData:
-        logger.debug(dct)
+        # logger.debug(dct)
         model = dct.pop("model")  # gets and deletes model from dictionary
+        logger.debug(model)
         start_stop: Optional[Dict] = dct.pop("start_stop", None)
         if start_stop is not None:
             startstop_obj = dto.StartStopData(**start_stop)
             dct["start_stop"] = startstop_obj
-        dev_class_str = f"Device{model.capitalize()}Data"
-        dev_class = get_class_from_dto(class_str=dev_class_str)
+        dev_class = get_device_data_class_from_str(model_name=model)
         return dev_class(**dct)
 
     @staticmethod
@@ -70,7 +73,7 @@ class DataclassJSONDecoder(json.JSONDecoder):
         data_nowcast = None
         if "data_nowcast" in dct:
             data_nowcast = dct["data_nowcast"]
-        return dto.TimeSeriesData(name, data, data_nowcast)
+        return dto.TimeSeriesData(id=name, data=data, data_nowcast=data_nowcast)
 
     def object_hook(self, dct):  # noqa
         carriers = []
@@ -110,11 +113,6 @@ class DataclassJSONDecoder(json.JSONDecoder):
             )
             return energy_system_data
         return dct
-
-    def default(self, obj: Any):
-        if isinstance(obj, dto.EnergySystemData):
-            return dto.EnergySystemData(obj=obj)  # noqa
-        return json.JSONEncoder.default(self, obj)  # noqa
 
 
 def serialize_oogeso_data(energy_system_data: dto.EnergySystemData):
@@ -159,13 +157,12 @@ class OogesoResultJSONDecoder(json.JSONDecoder):
             return result_data
         return dct
 
-    @staticmethod
-    def default(obj: Any):
-        if isinstance(obj, dto.SimulationResult):
-            return dto.SimulationResult()
-        return json.JSONEncoder().default(obj)
-
 
 def deserialize_oogeso_results(json_data):
     result_data = json.loads(json_data, cls=OogesoResultJSONDecoder)
     return result_data
+
+
+def serialize_oogeso_results(result_data: dto.SimulationResult):
+    json_string = json.dumps(result_data, cls=DataclassJSONEncoder, indent=2)
+    return json_string
