+++ tests/test_dto.py
@@ -1,4 +1,10 @@
 from oogeso import dto
+from oogeso.dto.serialisation import (
+    deserialize_oogeso_data,
+    deserialize_oogeso_results,
+    serialize_oogeso_data,
+    serialize_oogeso_results,
+)
 
 
 def test_create_energy_system_data():
@@ -76,16 +82,16 @@ def test_create_energy_system_data():
 
 def test_serialisation_deserialisation(testcase2_data, testcase2_expected_result):
 
-    data_str = serialisation.serialize_oogeso_data(testcase2_data)
+    data_str = serialize_oogeso_data(testcase2_data)
     assert isinstance(data_str, str)
 
-    data_obj = serialisation.deserialize_oogeso_data(data_str)
+    data_obj = deserialize_oogeso_data(data_str)
     assert isinstance(data_obj, dto.EnergySystemData)
 
     assert data_obj == testcase2_data
 
-    res_str = serialisation.serialize_oogeso_results(testcase2_expected_result)
+    res_str = serialize_oogeso_results(testcase2_expected_result)
     assert isinstance(res_str, str)
 
-    res_obj = serialisation.deserialize_oogeso_results(res_str)
+    res_obj = deserialize_oogeso_results(res_str)
     assert isinstance(res_obj, dto.SimulationResult)
